Add files via upload
This is a complete analysis addressing the demands of both the Machine Learning team and the Development team. This is a project, so the information can be made public. I worked with Pandas to analyze the dataset, create different DataFrames for the various requirements, and exported the new datasets for each team.